+++ src/main/scala/io/druid/indexer/spark/SparkBatchIndexTask.scala
@@ -19,29 +19,36 @@
 
 package io.druid.indexer.spark
 
-import java.io.{Closeable, File, IOException, PrintWriter}
-import java.nio.file.Files
-import java.util
-import java.util.{Objects, Properties}
-import com.fasterxml.jackson.annotation.{JsonCreator, JsonProperty}
-import com.google.common.base.{Preconditions, Strings}
+import com.fasterxml.jackson.annotation.JsonCreator
+import com.fasterxml.jackson.annotation.JsonProperty
+import com.google.common.base.Preconditions
+import com.google.common.base.Strings
 import com.google.common.collect.Iterables
 import com.google.common.io.Closer
 import com.metamx.common.logger.Logger
-import com.metamx.emitter.service.{ServiceEmitter, ServiceMetricEvent}
 import io.druid.common.utils.JodaUtils
 import io.druid.data.input.impl.ParseSpec
-import io.druid.indexing.common.actions.{LockTryAcquireAction, TaskActionClient}
-import io.druid.indexing.common.task.{AbstractTask, HadoopTask}
-import io.druid.indexing.common.{TaskStatus, TaskToolbox}
+import io.druid.indexing.common.actions.LockTryAcquireAction
+import io.druid.indexing.common.actions.TaskActionClient
+import io.druid.indexing.common.task.AbstractTask
+import io.druid.indexing.common.task.HadoopTask
+import io.druid.indexing.common.TaskStatus
+import io.druid.indexing.common.TaskToolbox
 import io.druid.java.util.common.granularity._
 import io.druid.query.aggregation.AggregatorFactory
 import io.druid.segment.IndexSpec
 import io.druid.segment.indexing.DataSchema
 import io.druid.timeline.DataSegment
-import org.apache.spark.scheduler.SparkListenerApplicationEnd
-import org.apache.spark.{SparkConf, SparkContext}
-import org.apache.spark.scheduler.{AccumulableInfo, SparkListener, SparkListenerStageCompleted}
+import java.io.Closeable
+import java.io.File
+impgitort java.io.IOException
+import java.io.PrintWriter
+import java.nio.file.Files
+import java.util
+import java.util.Objects
+import java.util.Properties
+import org.apache.spark.SparkConf
+import org.apache.spark.SparkContext
 import org.joda.time.Interval
 import scala.collection.JavaConversions._
 
@@ -400,43 +407,6 @@ object SparkBatchIndexTask
           sc.hadoopConfiguration.set(y, System.getProperty(x), "Druid Forking Property")
         }
       )
-
-      log.info("Initializing emitter for metrics")
-
-      lifecycle.start()
-
-      sc.addSparkListener(new SparkListener() {
-        // Emit metrics at the end of each stage
-        override def onStageCompleted(stageCompleted: SparkListenerStageCompleted): Unit = {
-          val dimensions = Map(
-            "taskId" -> task.getId,
-            "stageId" -> stageCompleted.stageInfo.stageId.toString,
-            "interval" -> SerializedJsonStatic.mapper.writeValueAsString(task.getIntervals)
-          )
-
-          val accumulatedInfo = stageCompleted.stageInfo.accumulables.toMap.flatMap {
-            case (_, AccumulableInfo(_, Some(name), _, Some(value: Long), _, _, _)) =>
-              Some(name -> value)
-
-            case _ =>
-              None
-          }
-
-          accumulatedInfo foreach { case (aggName, value) =>
-            log.debug("emitting metric: %s".format(aggName))
-            val eventBuilder = ServiceMetricEvent.builder()
-            dimensions foreach { case (n, v) => eventBuilder.setDimension(n, v)}
-            emitter.emit(
-              eventBuilder.build(aggName, value)
-            )
-          }
-        }
-
-        // Closes lifecycle when sc.stop() is called
-        override def onApplicationEnd(applicationEnd: SparkListenerApplicationEnd): Unit = {
-          lifecycle.stop()
-        }
-      })
 
       // Should be set by HadoopTask for job jars
       // Hadoop tasks use io.druid.indexer.JobHelper::setupClasspath to replicate their jars.
